Prevent unintended sign extension from int to long on 64-bit machines.
If the property format is 32, the property data will be stored as
an array of longs (which on 64-bit machines will be 64-bit values
that are padded in the upper 4 bytes).


git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@35455 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ gui/ui/widgets.c
@@ -18,6 +18,7 @@
 
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -100,7 +101,7 @@ static int gtkLoadIcon(GtkIconTheme *theme, gint size, GdkPixmap **gdkIcon, GdkB
             data = gdk_pixbuf_get_pixels(pixbuf);
 
             for (i = csize; i < guiIcon.collection_size; data += 4, i++)
-                guiIcon.collection[i] = (data[3] << 24) | AV_RB24(data);  // RGBA -> ARGB
+                guiIcon.collection[i] = (uint32_t)(data[3] << 24) | AV_RB24(data);  // RGBA -> ARGB
         }
 
         g_object_unref(pixbuf);
